Move helper to SignatureGenerator
