+++ Vonage/Request/ByteArrayToHexHelper.cs
@@ -1,33 +1,1 @@
 ﻿namespace Vonage.Request;
-
-internal static class ByteArrayToHexHelper
-{
-    ///// There is no built-in byte[] => hex string, so here's an implementation
-    /// http://stackoverflow.com/questions/311165/how-do-you-convert-byte-array-to-hexadecimal-string-and-vice-versa/24343727#24343727
-    /// We're not going to going with the unchecked version. Seems overkill for now.
-    internal static readonly uint[] _lookup32 = CreateLookup32();
-
-    internal static uint[] CreateLookup32()
-    {
-        var result = new uint[256];
-        for (var i = 0; i < 256; i++)
-        {
-            var s = i.ToString("X2");
-            result[i] = s[0] + ((uint)s[1] << 16);
-        }
-        return result;
-    }
-
-    internal static string ByteArrayToHex(byte[] bytes)
-    {
-        var lookup32 = _lookup32;
-        var result = new char[bytes.Length * 2];
-        for (var i = 0; i < bytes.Length; i++)
-        {
-            var val = lookup32[bytes[i]];
-            result[2 * i] = (char)val;
-            result[2 * i + 1] = (char)(val >> 16);
-        }
-        return new string(result);
-    }
-}
